+++ training_log_app/src/services/log_entry_service.py
@@ -1,0 +1,9 @@
+from entities.log_entry import LogEntry
+from entities.user import User
+from repositories.log_entry_repository import LogEntryRepository
+from repositories.user_repository import UserRepository
+
+
+class LogEntryService:
+    '''Class in charge of application logic'''
+    pass