+++ training_log_app/documentation/changelog.md
@@ -4,3 +4,4 @@
 - Added UserReposityory and LogEntryRepository classes, which are responsible for connections with database
 - Added database connection
 - Added configuration files for filesystem and database
+- User can add an entry to database